Reorganize Calculations (#30)
* add data_transformation folder

* add missing folder
