+++ src/vivarium_testing_utils/automated_validation/comparison.py
@@ -1,6 +1,8 @@
 import pandas as pd
 
-from vivarium_testing_utils.automated_validation.calculations import compute_metric
+from vivarium_testing_utils.automated_validation.data_transformation.calculations import (
+    compute_metric,
+)
 
 
 class Comparison:
